fix(shorebird_cli): provide exportOptionsPlist to all ipa build commands
